Update README.md
# Kickstarting with Excel

## Overview of Project

### Purpose
Louise's wants to be informed about how different campaigns performed in relation to their launch dates and funding goals. The Kickstarter dataset will help analyze and visualize the outcomes based on the Launch Date and Goals. This will help Louise make an informed decision and begin her Kickstarter preparations! 

## Analysis and Challenges

### Analysis of Outcomes Based on Launch Date
Analysis of Outcomes Based on Launch Date provides an overview of how Theaters perform in relations to different launch months. This is visualized via a pivot table and a line chart with the respective outcomes of "successful," "failed," and "canceled."

![Theater_Outcomes_vs_Launch.png]

### Analysis of Outcomes Based on Goals
Analysis of Outcomes Based on Goals provides an overview of how plays performed in respect to each monetary goal range. This is visualized via a pivot table and a line chart with the respective percentages for "successful," "failed," and "canceled" outcomes. 

![Outcomes_vs_Goals.png]

### Challenges and Difficulties Encountered
One challenge that I encountered when analyzing the "Outcome Based on Goals" was that when I tried to drag the function across a row, the function's range would change. For example when I dragged the function from "number of successful" to "number of failed" the criteria range would change from Kickstarter's D:D to F:F, which was not the correct criteria range for that specific filter. I was able to overcome this by just simply copying and pasting the function into the next cell then changing the criteria to filter for each respective "outcome." 

## Results

- What are two conclusions you can draw about the Outcomes based on Launch Date?
First conclusion is that projects launched in May have the highest percentage of successful rate. The second conclusion is that projects have the highest failed rate ranging in the months of May to July and October. 

- What can you conclude about the Outcomes based on Goals?
A conclusion that I can draw from the "Outcomes based on Goals" is that the "number successful" and "number failed" drastically decreases when the goal changes from $10,000 - $14999 to $15,000 - $19,999. 

- What are some limitations of this dataset?
Some limitations of this dataset is for "Outcomes based on Goals," we neglected all the "live" performances. The live performances could have held most of the goals. Additionally, in the sheet "Theater Outcomes by Launch Date" does not accurately represent the data for "plays." The parent category "Theater" encompasses all musical, plays, and spaces.   

- What are some other possible tables and/or graphs that we could create?
One possible table and graph that we can create is a "plays outcomes by launch date" that solely filters for the outcome of plays in relation to the launch months. This will give Louise a more accurate representation of specifically how "plays" throughout the years.